Split webext parse & stringify functions into their own files
