+++ js/src/index.ts
@@ -21,7 +21,8 @@ import { fluentSerializePattern } from './fluent-serialize.ts'
 import { mf2ParsePattern } from './mf2-parse.ts'
 import { mf2SerializePattern } from './mf2-serialize.ts'
 import type { Message, Pattern } from './model.ts'
-import { webextParsePattern, webextSerializePattern } from './webext.ts'
+import { webextParsePattern } from './webext-parse.ts'
+import { webextSerializePattern } from './webext-serialize.ts'
 import { xliffParsePattern } from './xliff-parse.ts'
 import { xliffSerializePattern } from './xliff-serialize.ts'
