+++ js/src/webext-serialize.ts
@@ -1,0 +1,36 @@
+/* Copyright Mozilla Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { ERROR_RESULT, SerializeError } from './errors.ts'
+import type { Pattern } from './model.ts'
+
+export function webextSerializePattern(
+  pattern: Pattern,
+  onError: (error: SerializeError) => void
+): string {
+  let str = ''
+  for (const part of pattern) {
+    if (typeof part === 'string') {
+      str += part.replace(/\$+/g, '$$$&')
+    } else if ('$' in part) {
+      str += part.attr?.source ?? `$${part.$}$`
+    } else {
+      const error = `webext: Unsupported pattern part ${JSON.stringify(part)}`
+      onError(new SerializeError(error, str.length))
+      str += ERROR_RESULT
+    }
+  }
+  return str
+}
